Work on stubbing out new item support.
Will re-configure routing.

we will swap to a :type/:mod/:id route, loading the correct editor, and pulling the save handler up to the parent of the actual editor.
Editor's will take data in through a standard prop (data), and provide a onChange callback which returns a { url: ..., data:{...} } object.
EditorChrome will hook into onChange, setting dirty flag (to enable save) and handle sending save data back to client.
Each editor will instantiate a blank object for it's data if none is provided, and onChange will provide a save url if the data is valid to be saved.